apps/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.16
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.9.0,<3
browser-use-sdk>=2.0.15
google-genai>=1.0.0
langchain-google-genai>=2.0.0
Pillow>=10.0.0
# Optional: install browser-use + playwright for local agent fallback (no BROWSER_USE_API_KEY)
# browser-use==0.12.6
# playwright==1.49.1  (then run: playwright install chromium)
